move logging to common
