+++ src/app/common/index.ts
@@ -1,3 +1,4 @@
 export * from './ellipsis-title.directive';
 export * from './window-resize-tracker.service';
 export * from './not-found';
+export * from './logging';
