+++ src/app/tasks-processing/task-executor/fake-task-executor.ts
@@ -1,6 +1,6 @@
 import { ITaskExecutor } from './i-task-executor';
 import { Hub } from '../../lego-hub';
-import { ILogger } from '../../logging';
+import { ILogger } from '../../common';
 import { PortCommandTask } from '../../types';
 
 export class FakeTaskExecutor implements ITaskExecutor {
